Changed Functors/Lens folder topology
