+++ src/Functors/Lens/index.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * lens functions
+ *
+ * @package bingo-functional
+ * @author Lochemem Bruno Michael
+ * @license Apache-2.0
+ */
+
+require_once __DIR__ . '/Lens.php';
+require_once __DIR__ . '/LensKey.php';
+require_once __DIR__ . '/LensPath.php';
+require_once __DIR__ . '/Over.php';
+require_once __DIR__ . '/Set.php';
+require_once __DIR__ . '/View.php';
